Case Study Financial Consultant
This project is made for Secure Data Management Course. It's implemented
based on the paper of Executing SQL over Encrypted Data in the
Database-Service-Provider Model (H. Hacigumus, B. Iyer, C. Li, and S.
Mehrotra) with some modification in the authentication and the
encryption method. This project use Diffie Hellman Key Exchange-1024
with AES-128 encryption scheme